Welcome to the GitHub repository for the Tweetster Disaster Monitor Android application!
This repository is dedicated to the development of the Android component of Tweetster Disaster Monitor. Our goal is to create a powerful and user-friendly disaster monitoring and response tool that leverages the capabilities of Natural Language Processing (NLP) and geolocation technology.

Within this repository, you will find all the necessary code and resources specific to the Android platform. We have carefully structured the project to ensure modularity, readability, and scalability. The codebase is organized into logical modules, including NLP integration, geolocation services, user interface components, and data visualization.